backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
pdfminer.six==20231228
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.2.2
fpdf2==2.7.8
google-cloud-aiplatform==1.49.0
google-cloud-vision==3.10.2
google-cloud-bigquery==3.25.0
firebase-admin==6.5.0
google-cloud-storage==2.16.0
httpx==0.27.0
tenacity==8.2.3
pytest==8.2.2
python-dateutil==2.9.0.post0
